+++ src/main/scala/repose/ReposeMod.scala
@@ -2,6 +2,9 @@ package repose
 
 import farseek.FarseekBaseMod
 import net.minecraftforge.fml.common.Mod
+import net.minecraftforge.fml.common.Mod.EventHandler
+import net.minecraftforge.fml.common.event.FMLModIdMappingEvent
+import repose.block.FallingBlockExtensions
 import repose.config.ReposeConfig
 
 /** @author delvr */
@@ -9,4 +12,6 @@ import repose.config.ReposeConfig
 object ReposeMod extends FarseekBaseMod {
 
     lazy val configuration = Some(ReposeConfig)
+
+    @EventHandler def handle(event: FMLModIdMappingEvent): Unit = FallingBlockExtensions.setSpongeNeighborOverrides()
 }